@@ -30,14 +30,17 @@
     ],
     "require": {
         "php": ">=8.1",
-        "symfony/http-foundation": "^6.1",
+        "symfony/http-foundation": "^6.2.x-dev",
         "hyperf/exception-handler": "~3.0.0",
         "hyperf/paginator": "~3.0.0",
         "hyperf/elasticsearch": "~3.0.0",
         "robinvdvleuten/ulid": "^5.0",
-        "godruoyi/php-snowflake": "^2.1"
+        "godruoyi/php-snowflake": "^2.1",
+        "hyperf/db-connection": "^3.0.x-dev",
+        "hyperf/snowflake": "^3.0"
     },
     "require-dev": {
+        "roave/security-advisories": "dev-latest",
         "friendsofphp/php-cs-fixer": "v3.10.0",
         "phpunit/phpunit": "^9.4"
     },
