@@ -7,7 +7,7 @@ use Hyperf\Database\Model\Builder;
 
 abstract class QueryFilter
 {
-    protected $builder;
+    protected Builder $builder;
 
     public function __construct(protected array $data = [])
     {
@@ -18,7 +18,7 @@ abstract class QueryFilter
         return $this->data;
     }
 
-    public function apply(Builder $builder)
+    public function apply(Builder $builder): Builder
     {
         $this->builder = $builder;
 
@@ -33,28 +33,31 @@ abstract class QueryFilter
         return $this->builder;
     }
 
-    protected function getInstrBuilder(string $field, string $keyword)
+    protected function getInstrBuilder(string $field, string $keyword): Builder
     {
-        return $this->builder->whereRaw("INSTR(`{$field}`, ?) > 0", ["$keyword"]);
+        return $this->builder->whereRaw("INSTR(`{$field}`, ?) > 0", [(string)$keyword]);
     }
 
-    protected function getInstrBuilderRelation(string $relation, string $field, string $keyword)
+    protected function getInstrBuilderRelation(string $relation, string $field, string $keyword): Builder
     {
-        return $this->builder->whereHas("{$relation}", function ($query) use ($field, $keyword) {
-            $query->whereRaw("INSTR(`{$field}`, ?) > 0", ["{$keyword}"]);
+        return $this->builder->whereHas((string)($relation), function ($query) use ($field, $keyword) {
+            $query->whereRaw("INSTR(`{$field}`, ?) > 0", [(string)($keyword)]);
         });
     }
 
-    protected function getBuilderRelationEqual(string $relation, string $field, string $keyword)
+    protected function getBuilderRelationEqual(string $relation, string $field, string $keyword): Builder
     {
-        return $this->builder->whereHas("{$relation}", function ($query) use ($field, $keyword) {
-            $query->where("{$field}", '=', "{$keyword}");
+        return $this->builder->whereHas((string)($relation), function ($query) use ($field, $keyword) {
+            $query->where((string)($field), '=', (string)($keyword));
         });
     }
 
+    /**
+     * @throws \JsonException
+     */
     protected function sort($sortObject)
     {
-        if (!empty($sortItems = json_decode($sortObject, true))) {
+        if (!empty($sortItems = json_decode($sortObject, true, 512, JSON_THROW_ON_ERROR))) {
             foreach ($sortItems as $itemName => $itemValue) {
 //               if (method_exists($this, $itemName)) {
 //                   call_user_func_array([$this, $itemName], array_filter([$itemValue]));
